requirements.txt
tensorflow
pandas
numpy
networkx
matplotlib
h5py
pycairo

